unit tests for assets
